Add Zod schema support to the 'inputSchemaRequired' middleware
